Fixed bug #16287: Forward NOT removed after deleting user account.

+++ deleteaccount.php
@@ -82,6 +82,13 @@ if ($ref!=$_SERVER['SCRIPT_FILENAME']){
 					die (_("Database error"));
 				}
 
+				// Removing forwards
+				$query = "DELETE FROM virtual WHERE alias='".$_GET['username']."' AND username=''";
+				$result = $handle->query($query);
+				if (DB::isError($result)) {
+					die (_("Database error"));
+				}
+
 				$query = "DELETE FROM accountuser WHERE username='".$_GET['username']."'";
 				$result = $handle->query($query);
 				if (DB::isError($result)) {
